requirements.txt
PyQt5>=5.15
mss>=9.0
opencv-python-headless>=4.8
Pillow>=10.0
numpy>=1.24
pyinstaller>=6.0
pywin32>=306; sys_platform == 'win32'
